requirements.txt
aerich==0.7.2 ; python_version >= "3.12" and python_version < "4.0"
aiofiles==24.1.0 ; python_version >= "3.12" and python_version < "4.0"
aiogram==3.22.0 ; python_version >= "3.12" and python_version < "4.0"
aiohappyeyeballs==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.12.15 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.4.0 ; python_version >= "3.12" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.29.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==25.4.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.11.12 ; python_version >= "3.12" and python_version < "4.0"
click==8.3.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
dictdiffer==0.9.0 ; python_version >= "3.12" and python_version < "4.0"
dotenv==0.9.9 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.8.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.11 ; python_version >= "3.12" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
magic-filter==1.0.12 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.7.0 ; python_version >= "3.12" and python_version < "4.0"
propcache==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.11.10 ; python_version >= "3.12" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.2.1 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.12" and python_version < "4.0"
tomlkit==0.13.3 ; python_version >= "3.12" and python_version < "4.0"
tortoise-orm==0.20.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.22.0 ; python_version >= "3.12" and python_version < "4.0"
